requirements.txt
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.20.45
filelock==3.8.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.2
jeepney==0.8.0
jsonschema==4.15.0
keyring==23.9.0
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
tomlkit==0.11.4
urllib3==1.26.12
virtualenv==20.16.4
webencodings==0.5.1
zipp==3.8.1
